+++ pinecone/config/__init__.py
@@ -1,0 +1,2 @@
+from .config import Config
+from .pinecone_config import PineconeConfig
