+++ pinecone/data/grpc/config.py
@@ -1,0 +1,33 @@
+from .retry import RetryConfig
+from typing import NamedTuple, Optional, Dict
+
+
+class GRPCClientConfig(NamedTuple):
+    """
+    GRPC client configuration options.
+
+    :param secure: Whether to use encrypted protocol (SSL). defaults to True.
+    :type traceroute: bool, optional
+    :param timeout: defaults to 2 seconds. Fail if gateway doesn't receive response within timeout.
+    :type timeout: int, optional
+    :param conn_timeout: defaults to 1. Timeout to retry connection if gRPC is unavailable. 0 is no retry.
+    :type conn_timeout: int, optional
+    :param reuse_channel: Whether to reuse the same grpc channel for multiple requests
+    :type reuse_channel: bool, optional
+    :param retry_config: RetryConfig indicating how requests should be retried
+    :type retry_config: RetryConfig, optional
+    :param grpc_channel_options: A dict of gRPC channel arguments
+    :type grpc_channel_options: Dict[str, str]
+    """
+
+    secure: bool = True
+    timeout: int = 20
+    conn_timeout: int = 1
+    reuse_channel: bool = True
+    retry_config: Optional[RetryConfig] = None
+    grpc_channel_options: Dict[str, str] = None
+
+    @classmethod
+    def _from_dict(cls, kwargs: dict):
+        cls_kwargs = {kk: vv for kk, vv in kwargs.items() if kk in cls._fields}
+        return cls(**cls_kwargs)
