+++ pinecone/data/grpc/protos/vector_column_service_pb2_grpc.py
