+++ pinecone/deprecated/__init__.py
@@ -1,0 +1,1 @@
+from .legacy_utils import *
