+++ pinecone/utils/constants.py
@@ -1,10 +1,7 @@
 import os
 import enum
 
-from pinecone.utils import get_environment, get_version
-
-PARENT_LOGGER_NAME = "pinecone"
-DEFAULT_PARENT_LOGGER_LEVEL = "ERROR"
+from .environment import get_environment, get_version
 
 MAX_MSG_SIZE = 128 * 1024 * 1024
 
@@ -29,10 +26,6 @@ class NodeType(str, enum.Enum):
 PACKAGE_ENVIRONMENT = get_environment() or "development"
 CLIENT_VERSION = get_version()
 CLIENT_ID = f"python-client-{CLIENT_VERSION}"
-
-TCP_KEEPINTVL = 60  # Sec
-TCP_KEEPIDLE = 300  # Sec
-TCP_KEEPCNT = 4
 
 REQUIRED_VECTOR_FIELDS = {"id", "values"}
 OPTIONAL_VECTOR_FIELDS = {"sparse_values", "metadata"}
