+++ pinecone/utils/error_handling.py
@@ -3,19 +3,14 @@ from functools import wraps
 
 from urllib3.exceptions import MaxRetryError, ProtocolError
 
-from pinecone import Config
-
 def validate_and_convert_errors(func):
     @wraps(func)
     def inner_func(*args, **kwargs):
-        Config.validate()  # raises exceptions in case of invalid config
         try:
             return func(*args, **kwargs)
         except MaxRetryError as e:
             if isinstance(e.reason, ProtocolError):
-                raise ProtocolError(
-                    f"Failed to connect to {e.url}; did you specify the correct index name?"
-                ) from e
+                raise ProtocolError(f"Failed to connect to {e.url}; did you specify the correct index name?") from e
             else:
                 raise
         except ProtocolError as e:
