+++ tests/unit/test_config.py
@@ -1,209 +1,80 @@
 import pinecone
-from pinecone.config import Config
+from pinecone.exceptions import ApiKeyError, PineconeConfigurationError
+from pinecone.config import PineconeConfig
 from pinecone.core.client.configuration import Configuration as OpenApiConfiguration
 
 import pytest
-import tempfile
 import os
 
+class TestConfig:
+    @pytest.fixture(autouse=True)
+    def run_before_and_after_tests(tmpdir):
+        """Fixture to execute asserts before and after a test is run"""
 
-@pytest.fixture(autouse=True)
-def run_before_and_after_tests(tmpdir):
-    """Fixture to execute asserts before and after a test is run"""
+        # Defend against unexpected env vars. Since we clear these variables below
+        # after each test execution, these should only be raised if there is
+        # test pollution in the environment coming from some other test file/setup.
+        known_env_vars = ["PINECONE_API_KEY", "PINECONE_ENVIRONMENT", "PINECONE_CONTROLLER_HOST"]
+        for var in known_env_vars:
+            if os.getenv(var):
+                raise ValueError(f"Unexpected env var {var} found in environment. Check for test pollution.")
 
-    # Defend against unexpected env vars. Since we clear these variables below
-    # after each test execution, these should only be raised if there is
-    # test pollution in the environment coming from some other test file/setup.
-    known_env_vars = ["PINECONE_API_KEY", "PINECONE_ENVIRONMENT", "PINECONE_CONTROLLER_HOST"]
-    for var in known_env_vars:
-        if os.getenv(var):
-            raise ValueError(f"Unexpected env var {var} found in environment. Check for test pollution.")
+        yield  # this is where the testing happens
 
-    # Unfortunately since config is a singleton, we need to reset it manually between tests
-    pinecone.init()
+        # Teardown : Unset any env vars created during test execution
+        for var in known_env_vars:
+            if os.getenv(var):
+                del os.environ[var]
 
-    yield  # this is where the testing happens
+    def test_init_with_environment_vars(self):
+        os.environ["PINECONE_API_KEY"] = "test-api-key"
+        os.environ["PINECONE_CONTROLLER_HOST"] = "test-controller-host"
 
-    # Teardown : Unset any env vars created during test execution
-    for var in known_env_vars:
-        if os.getenv(var):
-            del os.environ[var]
+        config = PineconeConfig()
 
+        assert config.API_KEY == "test-api-key"
+        assert config.HOST == "test-controller-host"
 
-def test_default_config():
-    """
-    Test that default config is loaded when no config is specified. This not really a valid config that can be used,
-    but adding this test just to document the legacy behavior.
-    """
-    pinecone.init()
-    assert Config.API_KEY == ""
-    assert Config.ENVIRONMENT == "us-west1-gcp"
-    assert Config.CONTROLLER_HOST == "https://controller.us-west1-gcp.pinecone.io"
-    assert Config.LOG_LEVEL == "ERROR"
+    def test_init_with_positional_args(self):
+        api_key = "my-api-key"
+        host = "my-controller-host"
 
+        config = PineconeConfig(api_key, host)
 
-def test_init_with_environment_vars():
-    os.environ["PINECONE_ENVIRONMENT"] = "test-env"
-    os.environ["PINECONE_API_KEY"] = "test-api-key"
-    os.environ["PINECONE_CONTROLLER_HOST"] = "test-controller-host"
+        assert config.API_KEY == api_key
+        assert config.HOST == host
 
-    pinecone.init()
+    def test_init_with_kwargs(self):
+        api_key = "my-api-key"
+        controller_host = "my-controller-host"
+        openapi_config = OpenApiConfiguration(api_key="openapi-api-key")
 
-    assert Config.API_KEY == "test-api-key"
-    assert Config.ENVIRONMENT == "test-env"
-    assert Config.CONTROLLER_HOST == "test-controller-host"
+        config = PineconeConfig(api_key=api_key, host=controller_host, openapi_config=openapi_config)
 
+        assert config.API_KEY == api_key
+        assert config.HOST == controller_host
+        assert config.OPENAPI_CONFIG == openapi_config
 
-def test_init_with_positional_args():
-    api_key = "my-api-key"
-    environment = "test-env"
-    host = "my-controller-host"
-    log_level = None  # deprecated property but still in positional list
-    openapi_config = OpenApiConfiguration(api_key="openapi-api-key")
+    def test_init_with_mispelled_kwargs(self, caplog):
+        PineconeConfig(api_key='my-api-key', unknown_kwarg='bogus')
+        assert "__init__ had unexpected keyword argument(s): unknown_kwarg" in caplog.text
 
-    pinecone.init(api_key, host, environment, log_level, openapi_config)
-
-    assert Config.API_KEY == api_key
-    assert Config.ENVIRONMENT == environment
-    assert Config.CONTROLLER_HOST == host
-    assert Config.OPENAPI_CONFIG == openapi_config
-
-
-def test_init_with_kwargs():
-    env = "test-env"
-    api_key = "my-api-key"
-    controller_host = "my-controller-host"
-    openapi_config = OpenApiConfiguration(api_key="openapi-api-key")
-
-    pinecone.init(api_key=api_key, environment=env, host=controller_host, openapi_config=openapi_config)
-
-    assert Config.API_KEY == api_key
-    assert Config.ENVIRONMENT == env
-    assert Config.CONTROLLER_HOST == controller_host
-    assert Config.OPENAPI_CONFIG == openapi_config
-
-
-def test_init_with_mispelled_kwargs(caplog):
-    pinecone.init(invalid_kwarg="value")
-    assert "init had unexpected keyword argument(s): invalid_kwarg" in caplog.text
-
-
-def test_init_with_file_based_configuration():
-    """Test that config can be loaded from a file"""
-    env = "ini-test-env"
-    api_key = "ini-api-key"
-    controller_host = "ini-controller-host"
-
-    with tempfile.NamedTemporaryFile(mode="w") as f:
-        f.write(
-            f"""
-        [default]
-        environment: {env}
-        api_key: {api_key}
-        controller_host: {controller_host}
+    def test_resolution_order_kwargs_over_env_vars(self):
         """
-        )
-        f.flush()
-
-        pinecone.init(config=f.name)
-
-    assert Config.API_KEY == api_key
-    assert Config.ENVIRONMENT == env
-    assert Config.CONTROLLER_HOST == controller_host
-
-
-def test_resolution_order_kwargs_over_env_vars():
-    """
-    Test that when config is present from multiple sources,
-    the order of precedence is kwargs > env vars
-    """
-    os.environ["PINECONE_ENVIRONMENT"] = "env-var-env"
-    os.environ["PINECONE_API_KEY"] = "env-var-api-key"
-    os.environ["PINECONE_CONTROLLER_HOST"] = "env-var-controller-host"
-
-    env = "kwargs-env"
-    api_key = "kwargs-api-key"
-    controller_host = "kwargs-controller-host"
-
-    pinecone.init(environment=env, api_key=api_key, host=controller_host)
-
-    assert Config.API_KEY == api_key
-    assert Config.ENVIRONMENT == env
-    assert Config.CONTROLLER_HOST == controller_host
-
-
-def test_resolution_order_kwargs_over_config_file():
-    """
-    Test that when config is present from multiple sources, the order of
-    precedence is kwargs > config file
-    """
-    env = "ini-test-env"
-    api_key = "ini-api-key"
-    controller_host = "ini-controller-host"
-
-    kwargs_api_key = "kwargs-api-key"
-
-    with tempfile.NamedTemporaryFile(mode="w") as f:
-        f.write(
-            f"""
-        [default]
-        environment: {env}
-        api_key: {api_key}
-        controller_host: {controller_host}
+        Test that when config is present from multiple sources,
+        the order of precedence is kwargs > env vars
         """
-        )
-        f.flush()
+        os.environ["PINECONE_API_KEY"] = "env-var-api-key"
+        os.environ["PINECONE_CONTROLLER_HOST"] = "env-var-controller-host"
 
-        pinecone.init(api_key=kwargs_api_key, config=f.name)
+        api_key = "kwargs-api-key"
+        controller_host = "kwargs-controller-host"
 
-    # Properties passed as kwargs take precedence over config file
-    assert Config.API_KEY == kwargs_api_key
+        config = PineconeConfig(api_key=api_key, host=controller_host)
 
-    # Properties not passed as kwargs loaded from config file
-    assert Config.ENVIRONMENT == env
-    assert Config.CONTROLLER_HOST == controller_host
+        assert config.API_KEY == api_key
+        assert config.HOST == controller_host
 
-
-def test_resolution_order_env_vars_over_config_file():
-    """
-    Test that when config is present from multiple sources, the order of precedence is
-    env vars > config file
-    """
-
-    os.environ["PINECONE_ENVIRONMENT"] = "env-var-env"
-    os.environ["PINECONE_API_KEY"] = "env-var-api-key"
-    os.environ["PINECONE_CONTROLLER_HOST"] = "env-var-controller-host"
-
-    with tempfile.NamedTemporaryFile(mode="w") as f:
-        f.write(
-            f"""
-        [default]
-        environment: ini-test-env
-        api_key: ini-api-key
-        controller_host: ini-controller-host
-        """
-        )
-        f.flush()
-
-        pinecone.init(config=f.name)
-
-    assert Config.API_KEY == "env-var-api-key"
-    assert Config.ENVIRONMENT == "env-var-env"
-    assert Config.CONTROLLER_HOST == "env-var-controller-host"
-
-
-def test_init_from_mixed_sources():
-    """
-    Test that even when some vars are found in a higher precedence source, the rest
-    are still loaded from lower precedence sources
-    """
-
-    os.environ["PINECONE_ENVIRONMENT"] = "env-var-env"
-    os.environ["PINECONE_API_KEY"] = "env-var-api-key"
-    controller_host = "kwargs-controller-host"
-
-    pinecone.init(host=controller_host)
-
-    assert Config.API_KEY == "env-var-api-key"
-    assert Config.ENVIRONMENT == "env-var-env"
-    assert Config.CONTROLLER_HOST == controller_host
+    def test_errors_when_no_api_key_is_present(self):
+        with pytest.raises(PineconeConfigurationError):
+            PineconeConfig()
